Refactor geo validation to validation.go file
